casi todo terminado, faltan los proveedores

+++ src/main/java/com/GenMan/GenMan/DTO/Relaciones/MateriaPrimaCantidadDTO.java
@@ -14,5 +14,5 @@ public class MateriaPrimaCantidadDTO {
     private Double precio;
     private String unidad;
     private Double cantidadDisponible;
-    private Integer cantidadRelacionada;
+    private Double cantidadRelacionada;
 }
