+++ src/main/java/com/GenMan/GenMan/DTO/Relaciones/ProductoCantidadDTO.java
@@ -14,5 +14,5 @@ public class ProductoCantidadDTO {
     private Double precio;
     private String unidad;
     private Double cantidadDisponible;
-    private Integer cantidadRelacionada;
+    private Double cantidadRelacionada;
 }
