+++ src/main/java/com/GenMan/GenMan/Entities/TablasIntermedias/Producto_Venta.java
@@ -32,7 +32,7 @@ public class Producto_Venta {
     private Producto producto;
 
     @Column(nullable = false)
-    private Integer cantidad;
+    private double cantidad;
 
     // Clase para clave compuesta PK
     @Data
